+++ examples_app/run_all_demos.sh
@@ -3,5 +3,5 @@ cargo build --release
 
 for file in src/bin/*.rs; do
     demo=$(basename "$file" .rs)
-    cargo run --bin "$demo"
+    cargo run --release --bin "$demo"
 done